requirement.txt
Django==2.0.10
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-shortuuidfield==0.1.3
djangorestframework==3.9.1
mysqlclient==1.4.1
Pillow==5.4.0
python-memcached==1.59
pytz==2018.7
shortuuid==0.5.0
six==1.12.0
